# perf: get all file data at once when downloading private file